Circumvent -Wformat-overflow in sprintf
In the hiscore phase of the title screen, the compiler can't be sure
that the string it is given is actually limited to 19 chars, so devise a
workaround to get rid of a warning. As long as the code in checkhiscore
is kept intact, the string is never over 19 chars and is always
null-terminated. The same sprintf code in that function doesn't cause a
warning.