Add a dockerized build script, update README
Use Ubuntu 14.04, as it has the oldest GLIBC version (2.19), being also
the oldest still-supported Ubuntu version. Its end of life comes soon,
so it needs to be updated by April 2024.

GLIBC versions are backwards-compatible, so using an older GLIBC version
means better compatibility. However, creating a build environment with a
very old GLIBC is worksome. Manually it would require building the whole
build toolchain. Using an out-of-the-box old GLIBC is limited to finding
a Docker base image that both has an old GLIBC and is not at end-of-life
so it still has package repositories available.